move HashAnnouncer.py to extras
