+++ tests/unit/dht/test_hash_announcer.py
@@ -1,7 +1,7 @@
 from twisted.trial import unittest
 from twisted.internet import defer, task
 from lbrynet.p2p import utils
-from lbrynet.dht.hashannouncer import DHTHashAnnouncer
+from lbrynet.extras.daemon.HashAnnouncer import DHTHashAnnouncer
 from tests.util import random_lbry_hash
 from tests.mocks import mock_conf_settings
